+++ lib/src/annotations/annotation_tool_variant.dart
@@ -8,5 +8,4 @@ enum AnnotationToolVariant {
   image,
   highlight,
   underline,
-  
 }